이제경 파일 이름 수정

+++ app/src/main/java/kr/ac/kopo/tripforu/Retrofit/GetRecommend.java
@@ -13,7 +13,7 @@ public class GetRecommend {
     @SerializedName("ScheduleId")
     private int scheduleId;
     @SerializedName("OwnerId")
-    private int ownerId;
+    private long ownerId;
     @SerializedName("Rating")
     private double rating;
     @SerializedName("Likes")
